+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Home from "./components/home";
+import Home from "./pages/home";
 import Panel from "./components/panel";
 
 function App() {
