+++ src/components/transactionsAnalysisCard.tsx
@@ -1,0 +1,41 @@
+import { icons } from "../assets";
+import { Heading } from "../helpers";
+import { ITransaction } from "../interface";
+
+const TransactionsAnalysisCard = ({
+  transactions,
+}: {
+  transactions: ITransaction[];
+}) => {
+  return (
+    <div className="w-1/2 h-72 p-4 rounded-lg">
+      <Heading header={"Transactions"} action={"today"} icon={icons.lens} />
+      {transactions.map((transaction: ITransaction, index: number) => (
+        <div
+          className="flex-center !justify-between gap-4 py-1"
+          style={{ borderTop: index !== 0 ? "1px solid black" : "" }}
+        >
+          <img
+            className="w-6 h-6"
+            src={icons.lens.icon}
+            alt={icons.lens.label}
+          />
+          <div className="flex-grow">
+            <p>{transaction.endUser}</p>
+            <p className="text-gray-300">{transaction.type}</p>
+          </div>
+          <div>
+            {transaction.increase ? (
+              <p className="text-green-600">+${transaction.amount}</p>
+            ) : (
+              <p className="text-red-600">-${transaction.amount}</p>
+            )}
+            <p className="text-gray-300">Debit Card</p>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default TransactionsAnalysisCard;
